refactored shadowBox and ProductDescription styles
