+++ src/components/ciSelectPlan/ThreeButtonsSelect.js
@@ -3,10 +3,11 @@ import { PseudoElement } from "../stdPlanOptions/PseudoElement";
 import { THEME } from "../../styles/theme";
 import { FONTS } from "../../styles/fonts";
 import SelectedIconSvg from "../../assets/icons/ciSelectPlan/selectedIcon.svg";
+import { shadowStyles } from "../../styles/shadowStyles";
 
 export const ThreeButtonsSelect = () => {
   return (
-    <View style={styles.sectorWrapper}>
+    <View style={[styles.sectorWrapper, shadowStyles.boxShadow]}>
       <Text style={styles.questionTitle}>Are you actively at work?</Text>
       <PseudoElement />
       <View style={styles.radioButtonsWrapper}>
@@ -40,10 +41,6 @@ export const ThreeButtonsSelect = () => {
 const styles = StyleSheet.create({
   sectorWrapper: {
     backgroundColor: THEME.COLOR.WHITE,
-    shadowColor: THEME.BOX_SHADOW.COLOR_BLACK,
-    shadowOffset: THEME.BOX_SHADOW.OFFSET,
-    shadowOpacity: THEME.BOX_SHADOW.OPACITY,
-    shadowRadius: THEME.BOX_SHADOW.SHADOW_RADIUS,
     borderRadius: 20,
     paddingHorizontal: 15,
     paddingTop: 19,
