+++ src/components/familyInfoPage/FamilyInfoItemStyles.js
@@ -13,10 +13,6 @@ export default StyleSheet.create({
 
   infoWrapper: {
     backgroundColor: THEME.COLOR.WHITE,
-    shadowColor: THEME.BOX_SHADOW.COLOR_BLACK,
-    shadowOpacity: THEME.BOX_SHADOW.OPACITY,
-    shadowOffset: THEME.BOX_SHADOW.OFFSET,
-    shadowRadius: 30,
     borderRadius: 20,
     paddingHorizontal: 23,
     paddingTop: 16,
