+++ src/components/homePage/HomePageTop.js
@@ -6,7 +6,6 @@ import { shadowStyles } from "../../styles/shadowStyles";
 
 export const HomePageTop = () => {
   return (
-    // TODO: implement external shadow styles
     <View style={[styles.wrapper, shadowStyles.boxShadow]}>
       <View>
         <Image
@@ -63,13 +62,6 @@ const styles = StyleSheet.create({
     borderRadius: 10,
     marginBottom: 9,
   },
-// TODO: implement external shadow styles
-  // shadowProp: {
-  //   shadowColor: THEME.BOX_SHADOW.COLOR_BLACK,
-  //   shadowOffset: THEME.BOX_SHADOW.OFFSET,
-  //   shadowOpacity: THEME.BOX_SHADOW.OPACITY,
-  //   shadowRadius: THEME.BOX_SHADOW.SHADOW_RADIUS,
-  // },
 
   lastTextEl: {
     marginBottom: 0,
