+++ src/components/personalInfoPage/TopPersonalInfo.js
@@ -1,10 +1,11 @@
 import { View, Text, StyleSheet } from "react-native";
 import { THEME } from "../../styles/theme";
 import { FONTS } from "../../styles/fonts";
+import { shadowStyles } from "../../styles/shadowStyles";
 
 export const TopPersonalInfo = () => {
   return (
-    <View style={styles.wrapper}>
+    <View style={[styles.wrapper, shadowStyles.boxShadow]}>
       <Text style={styles.text}>
         Review to confirm employee and family information to enroll in plans
         that best fit your needs
@@ -22,10 +23,6 @@ const styles = StyleSheet.create({
     paddingLeft: 20,
     borderBottomRightRadius: 20,
     borderBottomLeftRadius: 20,
-    shadowColor: THEME.BOX_SHADOW.COLOR_BLACK,
-    shadowOpacity: THEME.BOX_SHADOW.OPACITY,
-    shadowOffset: THEME.BOX_SHADOW.OFFSET,
-    shadowRadius: 30,
     marginBottom: 20,
   },
