+++ src/components/utils/PlanSectorStyles.js
@@ -7,10 +7,6 @@ export default StyleSheet.create({
     padding: 11,
     paddingBottom: 22,
     backgroundColor: THEME.COLOR.WHITE,
-    shadowColor: THEME.BOX_SHADOW.COLOR_BLACK,
-    shadowOffset: THEME.BOX_SHADOW.OFFSET,
-    shadowOpacity: THEME.BOX_SHADOW.OPACITY,
-    shadowRadius: THEME.BOX_SHADOW.SHADOW_RADIUS,
     borderRadius: 20,
     marginBottom: 16,
     borderWidth: 1,
